requirements.txt
python-dotenv
streamlit
os
ollama
OpenAI
markdown
